backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
flask-socketio==5.3.4
eventlet==0.33.3
werkzeug==2.2.3
itsdangerous==2.1.2
psycopg2==2.9.3
openai==0.27.8
python-dotenv==1.0.0
langchain==0.0.284
